# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MyMaestro",
    "version": "0.1.1"
  },
  "paths": {
...
    "/api/projets": {
      "get": {
        "tags": [
          "projets"
        ],
        "summary": "Lister Projets",
        "operationId": "lister_projets_api_projets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResumeProjet"
                  },
                  "type": "array",
                  "title": "Response Lister Projets Api Projets Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "director"
        ],
        "summary": "Creer Projet",
        "operationId": "creer_projet_api_projets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjetEntree"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Projet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projets/{projet_id}/chanson": {
      "put": {
        "tags": [
          "director"
        ],
        "summary": "Televerser Chanson",
        "description": "Chanson en corps brut ; sa durée est lue par ffprobe, un fichier illisible est refusé et supprimé.",
        "operationId": "televerser_chanson_api_projets__projet_id__chanson_put",
        "parameters": [
          {
            "name": "projet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Projet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projets/{projet_id}/director": {
      "get": {
        "tags": [
          "director"
        ],
        "summary": "Lire Etat Director",
        "operationId": "lire_etat_director_api_projets__projet_id__director_get",
        "parameters": [
          {
            "name": "projet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EtatDirector"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projets/{projet_id}/phases/{phase}/lancer": {
      "post": {
        "tags": [
          "director"
        ],
        "summary": "Lancer Phase",
        "operationId": "lancer_phase_api_projets__projet_id__phases__phase__lancer_post",
        "parameters": [
          {
            "name": "projet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projet Id"
            }
          },
          {
            "name": "phase",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Phase"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Projet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projets/{projet_id}/phases/{phase}/valider": {
      "post": {
        "tags": [
          "director"
        ],
        "summary": "Valider Phase",
        "operationId": "valider_phase_api_projets__projet_id__phases__phase__valider_post",
        "parameters": [
          {
            "name": "projet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projet Id"
            }
          },
          {
            "name": "phase",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Phase"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Projet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projets/{projet_id}/ecriture/brief": {
      "put": {
        "tags": [
          "director"
        ],
        "summary": "Definir Brief",
        "operationId": "definir_brief_api_projets__projet_id__ecriture_brief_put",
        "parameters": [
          {
            "name": "projet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Brief-Input"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EtatDirector"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projets/{projet_id}/ecriture/concepts": {
      "post": {
        "tags": [
          "director"
        ],
        "summary": "Proposer Concepts",
        "operationId": "proposer_concepts_api_projets__projet_id__ecriture_concepts_post",
        "parameters": [
          {
            "name": "projet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EtatDirector"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projets/{projet_id}/ecriture/concepts/{indice}/retenir": {
      "post": {
        "tags": [
          "director"
        ],
        "summary": "Retenir Concept",
        "operationId": "retenir_concept_api_projets__projet_id__ecriture_concepts__indice__retenir_post",
        "parameters": [
          {
            "name": "projet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projet Id"
            }
          },
          {
            "name": "indice",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Indice"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EtatDirector"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projets/{projet_id}/ecriture/chat": {
      "post": {
        "tags": [
          "director"
        ],
        "summary": "Envoyer Message",
        "operationId": "envoyer_message_api_projets__projet_id__ecriture_chat_post",
        "parameters": [
          {
            "name": "projet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageEntree"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EtatDirector"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projets/{projet_id}/ecriture/decoupage": {
      "post": {
        "tags": [
          "director"
        ],
        "summary": "Ecrire Decoupage",
        "operationId": "ecrire_decoupage_api_projets__projet_id__ecriture_decoupage_post",
        "parameters": [
          {
            "name": "projet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EtatDirector"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projets/{projet_id}/analyse/lignes": {
      "put": {
        "tags": [
          "director"
        ],
        "summary": "Modifier Lignes",
        "operationId": "modifier_lignes_api_projets__projet_id__analyse_lignes_put",
        "parameters": [
          {
            "name": "projet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/LigneParoles"
                },
                "title": "Lignes"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EtatDirector"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projets/{projet_id}/compatibilites": {
      "get": {
        "tags": [
          "director"
        ],
        "summary": "Lire Compatibilites",
        "operationId": "lire_compatibilites_api_projets__projet_id__compatibilites_get",
        "parameters": [
          {
            "name": "projet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/CompatibiliteMoteur"
                    }
                  },
                  "title": "Response Lire Compatibilites Api Projets  Projet Id  Compatibilites Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projets/{projet_id}/casting": {
      "put": {
        "tags": [
          "director"
        ],
        "summary": "Modifier Casting",
        "operationId": "modifier_casting_api_projets__projet_id__casting_put",
        "parameters": [
          {
            "name": "projet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CastingModification"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Projet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projets/{projet_id}/plans/{plan_id}": {
      "patch": {
        "tags": [
          "director"
        ],
        "summary": "Modifier Plan",
        "operationId": "modifier_plan_api_projets__projet_id__plans__plan_id__patch",
        "parameters": [
          {
            "name": "projet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projet Id"
            }
          },
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanModification"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageEntree": {
        "properties": {
          "texte": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Texte"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "texte"
        ],
        "title": "MessageEntree"
      },
...
      "ResumeProjet": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "titre": {
            "type": "string",
            "title": "Titre"
          },
          "module": {
            "type": "string",
            "title": "Module"
          },
          "format": {
            "$ref": "#/components/schemas/FormatImage"
          },
          "cree_le": {
            "type": "string",
            "title": "Cree Le"
          },
          "nb_plans": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Nb Plans"
          },
          "recette_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recette Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "titre",
          "module",
          "format",
          "cree_le",
          "nb_plans",
          "recette_id"
        ],
        "title": "ResumeProjet"
      },
...
    }
  }
}